Attempt to fix things

+++ Pad.cs
@@ -1,5 +1,6 @@
 using Godot;
 using System;
+using System.Threading.Tasks;
 
 public partial class Pad : Area2D
 {
@@ -26,7 +27,7 @@ public partial class Pad : Area2D
 	//[Signal] public delegate void OnPadCompleteEventHandler(bool success);
 	
 	
-	public void Initialize(float lifetime, float delay)
+	public async void Initialize(float lifetime, float delay)
 	{
 		GD.Print("Init pad in ", delay);
 		IsReady = false;
@@ -34,7 +35,12 @@ public partial class Pad : Area2D
 		
 		Sprite2D sprite = GetNode<Sprite2D>("Sprite2D");
 		sprite.Visible = false;
-		delay = 0;
+		//await ToSignal(GetTree().CreateTimer(delay), "timeout");
+		int ms = Mathf.RoundToInt(delay * 1000);
+		await Task.Delay(ms);
+		Enable();
+		return;
+		
 		if (delay == 0)
 		{
 			Enable();
@@ -46,23 +52,26 @@ public partial class Pad : Area2D
 		timer.Start(delay);
 	}
 	
-	private void Enable()
+	private async void Enable()
 	{
-		GD.Print("enable");
 		IsReady = true;
 		Sprite2D sprite = GetNode<Sprite2D>("Sprite2D");
 		Color randColor = Colors[GD.Randi() % Colors.Length];
-		GD.Print(randColor);
 		sprite.Modulate = randColor;
 		sprite.Visible = true;
 		
+		GD.Print("pad enabled for ", Lifetime);
+		int ms = Mathf.RoundToInt(Lifetime * 1000);
+		await Task.Delay(ms);
+		
+		//await ToSignal(GetTree().CreateTimer(Lifetime), "timeout");
+		OnTimerTimeout();
+		return;
 		//var timer = GetNode<Timer>("Timer");
 		timer.Timeout -= Enable;
 		timer.Timeout += OnTimerTimeout;
 		timer.Stop();
 		timer.Start(Lifetime);
-		GD.Print("pad enabled for ", Lifetime);
-
 	}
 
 	private void OnPadBodyEntered(Node2D body)
@@ -76,14 +85,16 @@ public partial class Pad : Area2D
 	
 	private void OnTimerTimeout()
 	{
-		GD.Print("ALARM");
-		Kill(false);
+
+			Kill(false);
 	}
 	
 	// also kill if it has been alive for too long
 	public void Kill(bool success)
 	{
 		// NEED TO TELL THE SEQUENE OF THE STATUS
+		if (!IsReady) return;
+		IsReady = false;
 		GD.Print("kill " + Name);
 		Sequence seq = GDExtensions.GetParentOfType<Sequence>(this);//  GetParent().GetNode<Sequence>("Sequence");
 		seq.PadComplete(success);
